feat(quotes): Sprint 9 — virgola decimali + DXF select all
═══ 9a — Virgola come decimal separator (bug "0,5 → 05")

Bug riportato: nel preventivo digitare 0,5 nelle ore mostra 05 (intero).
Causa: parseFloat("0,5") in JavaScript ritorna 0 perché si ferma alla
virgola. In Italia l'utente digita naturalmente 0,5 → tutto il preventivo
finisce con valori 0.

Fix: nuovo helper lib/decimalInput.ts con parseDecimal(s) che fa
.replace(',', '.') prima di parseFloat. 25 sostituzioni:
- PhaseEditor.tsx: 5 (setup_hours, cycle_hours, fixed_cost, ecc.)
- PartCard.tsx: 10 (raw_x/y/z/diameter, peso, costi, margine, prezzo)
- EdmPhaseFields.tsx: 2 (cut_length_mm, cut_height_mm)
- QuoteWizard.tsx: 1 (global_margin_percent)
- NewQuote2DPage.tsx: 3 (default_quantity, margin, drill_diameter)
- QuoteEditor.tsx: 4 (margine, transport, packaging, ecc.)

parseInt per quantity/n_components lasciato invariato (interi, no decimali).

═══ 9b — DXF profili: seleziona/deseleziona tutti

2 button compatti nell'header del Dxf2dProfileList:
- "Seleziona tutti" (disabled se già tutti selezionati)
- "Deseleziona tutti" (disabled se nessuno selezionato)

Callback onSelectAll(selected: boolean) propagata a NewQuote2DPage che
gestisce setSelectedIds via Set di tutti gli id o vuoto.

Verifiche: TS check 0 errori, parseFloat residui 0 nei file toccati.

+++ frontend/src/components/quotes/EdmPhaseFields.tsx
@@ -1,6 +1,7 @@
 import { Input } from '@/components/ui/input'
 import { Zap, Unlock } from 'lucide-react'
 import type { Phase, CuttingCycle } from '@/types'
+import { parseDecimal } from '@/lib/decimalInput'
 
 interface Props {
   phase: Phase
@@ -41,7 +42,7 @@ export default function EdmPhaseFields({ phase, edmAuto, cuttingCycles, onChange
           <Input type="number" step="1" min="0" className="mt-1 h-9 text-sm"
             value={phase.cut_length_mm ?? ''}
             placeholder="es. 320"
-            onChange={e => onChange('cut_length_mm', e.target.value === '' ? null : parseFloat(e.target.value))}
+            onChange={e => onChange('cut_length_mm', e.target.value === '' ? null : parseDecimal(e.target.value))}
             onBlur={onBlur} />
         </div>
         <div>
@@ -49,7 +50,7 @@ export default function EdmPhaseFields({ phase, edmAuto, cuttingCycles, onChange
           <Input type="number" step="0.5" min="0" className="mt-1 h-9 text-sm"
             value={phase.cut_height_mm ?? ''}
             placeholder="es. 40"
-            onChange={e => onChange('cut_height_mm', e.target.value === '' ? null : parseFloat(e.target.value))}
+            onChange={e => onChange('cut_height_mm', e.target.value === '' ? null : parseDecimal(e.target.value))}
             onBlur={onBlur} />
         </div>
         <div>
